Delete WidgetView
